Test of everything

+++ src/GameEngine/Assets/Animation.h
@@ -1,21 +1,19 @@
 #ifndef ANIMATION_H
 #define ANIMATION_H
 
-#include <cstddef>
-
 #include "SFML/Graphics/Texture.hpp"
 
 class Animation
 {
     const sf::Texture& _texture;
-    size_t _frameCount;
-    size_t _frameDuration;
+    short _frameCount;
+    int _frameDuration;
 
     sf::Vector2i _size;
 
 public:
     // ToDo:
-    Animation(const sf::Texture& texture, size_t frameCount, size_t frameDuration);
+    Animation(const sf::Texture& texture, short frameCount, int frameDuration);
 
     // ToDo:
     Animation(const Animation& other);
@@ -24,8 +22,8 @@ public:
     Animation(const Animation&& other);
 
     const sf::Texture& GetTexture() const { return _texture; }
-    size_t FrameCount() const { return _frameCount; }
-    size_t FrameDuration() const { return _frameDuration; }
+    short FrameCount() const { return _frameCount; }
+    int FrameDuration() const { return _frameDuration; }
     const sf::Vector2i& Size() const { return _size; }
 };
