+++ src/GameEngine/Assets/AssetManager.h
@@ -1,6 +1,5 @@
 #ifndef ASSETMANAGER_H
 #define ASSETMANAGER_H
-#include <cstddef>
 #include <string>
 #include <unordered_map>
 
@@ -18,7 +17,7 @@ public:
     void LoadFromFile(const std::string& path);
 
     void AddTexture(const std::string& name, const std::string& path);
-    void AddAnimation(const std::string& name, const std::string& textureName, size_t frameCount, size_t frameDuration);
+    void AddAnimation(const std::string& name, const std::string& textureName, short frameCount, int frameDuration);
     void AddFont(const std::string& name, const std::string& path);
 
     const sf::Texture& GetTexture(const std::string& name) const;
